+++ src/db/PinsCRUD.ts
@@ -19,14 +19,22 @@ export default class PinCRUD extends EventEmitter implements CRUD<Pin['channel']
         this.sequenceCRUD = sequenceCRUD
     }
 
-    insert = (obj: Pin) => this.db.insert(obj)
+    insert = async (obj: Pin) => {
+        const results = await this.db.insert(obj)
+        this.emit('insert', results)
+        return results
+    }
     update = async (id: Pin['channel'], obj: Partial<Pin>) => {
-        return await this.db.update(id, obj)
+        const results = await this.db.update(id, obj)
+        this.emit('update', results)
+        return results
     }
 
     set = this.update
     remove = async (key: Pin['channel']) => {
-        return await this.db.deleteByKey(key)
+        if (!await this.db.exists(key)) return
+        await this.db.deleteByKey(key)
+        this.emit('remove', key)
     }
     get = (key: Pin['channel']) => this.db.findByKey(key)
     list = () => this.db.findAll()
